+++ test/test_pytest.py
@@ -739,14 +739,9 @@ def test_health_check_failure(ctx, testdir, openapi3_base_url, snapshot):
         components={
             "schemas": {
                 "Item": {
-                    "additionalProperties": False,
-                    "required": [
-                        "A",
-                        "B",
-                        "C",
-                        "D",
-                    ],
-                    "type": "object",
+                    "type": "string",
+                    "format": "date-time",
+                    "pattern": "abc",
                 }
             }
         },